avia_admin/requirements.txt
Django==4.2.14
django-split-settings==1.3.2
psycopg==3.1.18
psycopg-binary==3.1.18
django-debug-toolbar==4.4.6
drf-spectacular==0.28.0
djangorestframework==3.15.2